Restrain GitHub api calls by the server
When a test is being viewed, the server makes some GitHub api
calls to potentially show some extra information to the user.
Most of the time the results of the api calls are cached.

This system works well under normal use but it seems to
break down somewhat when crawlers systematically view all
tests. In that case the server becomes just a proxy for
executing pointless GitHub api calls.

From a practical side this has not been a real problem.
When a crawler visits, the server seems to execute up to
500 GitHub api calls/hour which is far below the limit of
5000.

Still, the server executing pointless GitHub api calls on behalf
of random crawlers seems wrong, so in this PR we restrict this
as follows:

- Authenticated users are allowed to make api calls (a crawler
will be unauthenticated).
- Api calls are allowed for tests < 30 days old. These will
be served from the cache most of the time.
- In other cases api calls are not allowed.