Log failing API calls to event log.
Occasionally there are reports of GitHub API calls that
fail and then succeed when repeated. E.g.
https://github.com/official-stockfish/fishtest/pull/2293#issuecomment-2956858006

It would be interesting to know how often this occurs.
Therefore we now log such events in the event log.
We do this by saving certain error replies in the GH API cache
but we do not use these entries afterwards. So there is no
negative caching. Not at this time anyway.

It is not obvious how to test this PR in real life, but it
is tested in test_github_not_found() in test_github_api.py.